+++ jest.config.js
@@ -12,6 +12,7 @@ module.exports = {
   collectCoverageFrom: [
     "packages/ariakit/src/**/*.{js,ts,tsx}",
     "packages/ariakit-utils/src/**/*.{js,ts,tsx}",
+    "!**/__examples__/**",
     "!**/*test.{js,ts,tsx}",
     "!**/test-*.{js,ts}",
   ],