+++ packages/ariakit/src/combobox/__examples__/combobox-textarea/test-chrome.ts
@@ -1,6 +1,7 @@
 import { expect, test } from "@playwright/test";
 
-test("popover is positioned correctly", async ({ page }) => {
+test("popover is positioned correctly", async ({ page }, testInfo) => {
+  testInfo.snapshotSuffix = "";
   await page.goto("/examples/combobox-textarea");
 
   const textarea = await page.locator("role=combobox[name='Comment']");
@@ -10,14 +11,7 @@ test("popover is positioned correctly", async ({ page }) => {
   const popover = await page.locator(".popover[role='listbox']");
   await expect(popover).toBeVisible();
 
-  const { x, y } = (await textarea.boundingBox())!;
-
-  expect(await popover.boundingBox()).toEqual({
-    width: 180,
-    height: 186,
-    x: process.platform === "linux" ? x + 65 : x + 66,
-    y: y + 34,
-  });
+  expect(await textarea.screenshot()).toMatchSnapshot();
 
   await textarea.type("\n\n\n\n\n\n\n\n\n\n");
   await textarea.press("ArrowUp");
@@ -26,20 +20,10 @@ test("popover is positioned correctly", async ({ page }) => {
   await textarea.press("ArrowUp");
   await textarea.type("@");
 
-  expect(await popover.boundingBox()).toEqual({
-    width: 180,
-    height: 186,
-    x: x + 23,
-    y: y + 44,
-  });
+  expect(await textarea.screenshot()).toMatchSnapshot();
 
   await page.mouse.wheel(0, -50);
   await page.waitForTimeout(250);
 
-  expect(await popover.boundingBox()).toEqual({
-    width: 180,
-    height: 186,
-    x: x + 23,
-    y: y + 94,
-  });
+  expect(await textarea.screenshot()).toMatchSnapshot();
 });