+++ packages/ariakit/src/popover/popover-state.ts
@@ -81,6 +81,10 @@ function getAnchorElement(
   };
 }
 
+function isValidPlacement(flip: string): flip is Placement {
+  return /^(?:top|bottom|left|right)(?:-(?:start|end))?$/.test(flip);
+}
+
 /**
  * Provides state for the `Popover` components.
  * @example
@@ -152,9 +156,26 @@ export function usePopoverState({
           }),
         ];
 
-        if (flip) {
+        if (flip !== false) {
+          const fallbackPlacements =
+            typeof flip === "string" ? flip.split(" ") : undefined;
+
+          if (
+            fallbackPlacements !== undefined &&
+            !fallbackPlacements.every(isValidPlacement)
+          ) {
+            throw new Error(
+              "`flip` expects a spaced-delimited list of placements"
+            );
+          }
+
           // https://floating-ui.com/docs/flip
-          middleware.push(middlewares.flip({ padding: overflowPadding }));
+          middleware.push(
+            middlewares.flip({
+              padding: overflowPadding,
+              fallbackPlacements: fallbackPlacements,
+            })
+          );
         }
 
         if (slide || overlap) {
@@ -418,11 +439,27 @@ export type PopoverState = DialogState & {
    */
   shift: number;
   /**
-   * Whether the popover should flip to the opposite side of the viewport
+   * Controls the behavior of the popover when it overflows the viewport.
+   *
+   * If a boolean, specifies whether the popover should flip to the opposite side
    * when it overflows.
+   *
+   * If a string, indicates the preferred fallback placements when it overflows.
+   * The placements must be spaced-delimited, e.g. "top left".
+   *
+   * @example
+   *  ```jsx
+   *  const popover = usePopoverState({
+   *      placement: "right",
+   *      // In case the `right` placement overflows the viewport,
+   *      // the placement will fallback to the `bottom` or `top`,
+   *      // in that order, depending on where it fits.
+   *      flip: "bottom top",
+   *  });
+   * ```
    * @default true
    */
-  flip: boolean;
+  flip: boolean | string;
   /**
    * Whether the popover should slide when it overflows.
    * @default true