+++ packages/ariakit/CHANGELOG.md
@@ -1,5 +1,13 @@
 # ariakit
 
+## 2.0.0-next.27
+
+### Minor Changes
+
+- The `flip` prop on `usePopoverState` and derived hooks now supports a `string` of fallback placements. ([#1337](https://github.com/ariakit/ariakit/pull/1337))
+
+  See the [`popover-flip`](https://ariakit.org/examples/popover-flip) example.
+
 ## 2.0.0-next.26
 
 ### Minor Changes